Add "string" to cppRawDelimiter group name
This change fixes handling of raw strings in matchparen - a standard plugin that highlights matching parentheses.
The bug that I'm fixing manifests in the following code:

foo(bar(
  R"( )"
) );

If you move over parens in this code in Vim, you'll see that wrong ones are highlighted. This is because matchparen uses syntax highlighting to find (and skip) string literals. To be precise, it assumes that all string literals are highlighted with a syntax group whose name contains "string". Because we used "cppRawDelimiter", which does not contain "string" in it, parens in raw string delimiters confused the matchparen algorithm.